# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Scenario Lifecycle Management Proto App (barreleye-proto)
host: localhost:8088
# basePath prefixes all resource paths
basePath: /v1
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
################################################################################
#                                  Security                                    #
################################################################################
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
################################################################################
#                                  Paths                                       #
################################################################################
paths:
  /scenarios/single:
    x-swagger-router-controller : scenarios_single
    post:
      description: Create new singel scenario object
      operationId: addSingleScenario
      security:
        - api_key: []
      parameters:
        - in: body
          name: body
          description: Single Scenario object that needs to be added
          required: true
          schema:
            $ref: "#/definitions/SingleScenario"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SingleScenarioResponse"
        "400":
          description: Invalid scenario supplied
  /scenarios/single/{s_id}:
    x-swagger-router-controller : scenarios_single
    get:
      description: Returns 'Single scenario' object by id to the caller
      operationId: getSingleScenarioById
      parameters:
        - name: s_id
          in: path
          description: The id of the single scenario
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SingleScenarioResponse"
        "404":
          description: Scenario not found
        "400":
          description: Invalid scenario supplied
      security:
       - api_key: []
    put:
      description: Update 'Single scenario' object
      operationId: setSingleScenarioById
      parameters:
        - in: body
          name: body
          required: true
          description: The id of the single scenario
          schema:
            $ref: "#/definitions/SingleScenario"
        - in: path
          name: s_id
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SingleScenarioResponse"
        "404":
          description: Scenario not found
        "400":
          description: Invalid scenario supplied
      security:
       - api_key: []
  /swagger:
    x-swagger-pipe: swagger_raw
################################################################################
#                                 Definitions                                  #
################################################################################
definitions:
  SingleScenario:
    type: object
    required:
      - s_name
    properties:
      s_name:
        type: string
      s_type:
        type: string
      s_pattern:
        type: string
      s_apply_type:
        type: string
      s_run_period:
        type: string
      s_detect_action_policy:
        type: string
      s_profile_period:
        type: string
      s_datetime:
        type: string
  SingleScenarioResponse:
    type: object
    properties:
      s_id:
        type: integer
      s_name:
        type: string
      s_type:
        type: string
      s_pattern:
        type: string
      s_apply_type:
        type: string
      s_run_period:
        type: string
      s_detect_action_policy:
        type: string
      s_profile_period:
        type: string
      s_datetime:
        type: string
      s_created_time:
        description: Epoch time (ms)
        type: integer
        format: int64
  ErrorResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string